fix: move Genkit class to a library and export
Added docs to plugin library and Genkit class
